requirements.txt
flask==2.0.1
python-dotenv==0.19.2
google-generativeai==0.3.1
requests==2.26.0